修改单位转换方法和调用方式

+++ app/src/main/java/com/lodz/android/agiledevkt/modules/bottomsheet/ImgBottomSheetDialog.kt
@@ -35,7 +35,7 @@ class ImgBottomSheetDialog(context: Context) : BaseBottomSheetDialog(context) {
     override fun onBehaviorInit(behavior: BottomSheetBehavior<*>) {
         mBehavior = behavior
 
-        behavior.peekHeight = context.dp2px(200).toInt()
+        behavior.peekHeight = context.dp2px(200)
         behavior.setBottomSheetCallback(object : BottomSheetBehavior.BottomSheetCallback() {
             override fun onStateChanged(bottomSheet: View, newState: Int) {
                 if (!isUserDismiss) {
